Refactor: Organize backend routes into separate modules
- Create backend/routes/ directory structure
- Move integrations routes from app.py to routes/integrations.py
- Move basic routes (/, /api/health, /api/hello) from app.py to routes/basic.py
- Move projects routes from projects.py to routes/projects.py
- Create routes/conversations.py and move conversation endpoints from projects.py
- Update app.py to import and register all route blueprints
- Remove old projects.py file (functionality moved to routes/)
- Keep keys.py as utility module (still used by routes)

This improves code organization and maintainability by separating concerns
and grouping related functionality together.
