+++ backend/routes/__init__.py
@@ -1,0 +1,4 @@
+"""
+Routes package for OpenVibe backend.
+Contains all route blueprints organized by functionality.
+"""
