+++ backend/routes/projects.py
@@ -85,61 +85,6 @@ def save_projects(projects):
         logger.debug(f"💾 Error type: {type(e).__name__}")
         return False
 
-def load_conversations(project_id):
-    """Load conversations for a specific project from file"""
-    conversations_file = get_conversations_file(project_id)
-    logger.debug(f"📁 Loading conversations from: {conversations_file}")
-    logger.debug(f"📁 File exists: {conversations_file.exists()}")
-    
-    if conversations_file.exists():
-        try:
-            logger.debug(f"📁 File size: {conversations_file.stat().st_size} bytes")
-            
-            with open(conversations_file, 'r') as f:
-                content = f.read()
-                logger.debug(f"📁 File content length: {len(content)} characters")
-                
-                conversations = json.loads(content)
-                logger.info(f"📁 Successfully loaded {len(conversations)} conversations for project {project_id}")
-                return conversations
-        except (json.JSONDecodeError, IOError) as e:
-            logger.error(f"❌ Failed to load conversations for project {project_id}: {e}")
-            return []
-    else:
-        logger.info(f"📁 Conversations file doesn't exist for project {project_id}, returning empty list")
-    return []
-
-def save_conversations(project_id, conversations):
-    """Save conversations for a specific project to file"""
-    logger.debug(f"💾 Saving {len(conversations)} conversations for project {project_id}")
-    
-    try:
-        DATA_DIR.mkdir(parents=True, exist_ok=True)
-        logger.debug(f"💾 Data directory created/verified")
-    except Exception as e:
-        logger.error(f"❌ Failed to create data directory: {e}")
-        return False
-    
-    conversations_file = get_conversations_file(project_id)
-    logger.debug(f"💾 Conversations file path: {conversations_file}")
-    
-    try:
-        # Create backup if file exists
-        if conversations_file.exists():
-            backup_file = conversations_file.with_suffix('.json.backup')
-            logger.debug(f"💾 Creating backup: {backup_file}")
-            conversations_file.rename(backup_file)
-        
-        with open(conversations_file, 'w') as f:
-            json.dump(conversations, f, indent=2)
-        
-        logger.info(f"💾 Successfully saved {len(conversations)} conversations for project {project_id}")
-        logger.debug(f"💾 File size: {conversations_file.stat().st_size} bytes")
-        return True
-    except IOError as e:
-        logger.error(f"❌ Failed to save conversations for project {project_id}: {e}")
-        return False
-
 def create_slug(name):
     """Convert project name to slug format"""
     # Convert to lowercase and replace spaces/special chars with hyphens
@@ -467,7 +412,6 @@ def get_fly_status(project_slug, fly_token):
         logger.error(f"💥 Fly.io status check error: {str(e)}")
         return None
 
-
 def delete_github_repo(repo_url, github_token):
     """Delete a GitHub repository"""
     logger.info(f"🗑️ Deleting GitHub repo: {repo_url}")
@@ -515,7 +459,6 @@ def delete_github_repo(repo_url, github_token):
     except Exception as e:
         logger.error(f"💥 Error deleting GitHub repo: {str(e)}")
         return False, f"Error deleting repository: {str(e)}"
-
 
 def delete_fly_app(app_name, fly_token):
     """Delete a Fly.io app"""
@@ -568,7 +511,6 @@ def delete_fly_app(app_name, fly_token):
     except Exception as e:
         logger.error(f"💥 Error deleting Fly.io app: {str(e)}")
         return False, f"Error deleting app: {str(e)}"
-
 
 def create_github_repo(repo_name, github_token, fly_token):
     """Create a GitHub repository from template and set FLY_API_TOKEN secret"""
@@ -781,147 +723,123 @@ def get_project(slug):
                 
                 # Get GitHub status if token is available
                 if github_token and project.get('github_url'):
-                    logger.info(f"🔍 Checking GitHub status for {project['name']}")
                     github_status = get_github_status(project['github_url'], github_token)
                 
                 # Get Fly.io status if token is available
-                if fly_token:
-                    logger.info(f"🔍 Checking Fly.io status for {project['name']}")
+                if fly_token and project.get('slug'):
                     fly_status = get_fly_status(project['slug'], fly_token)
                     
             except Exception as e:
                 logger.warning(f"⚠️ Error getting status information: {str(e)}")
-                # Continue without status info
         
         # Add status information to project
         project_with_status = project.copy()
-        project_with_status['github_status'] = github_status
-        project_with_status['fly_status'] = fly_status
+        if github_status:
+            project_with_status['github_status'] = github_status
+        if fly_status:
+            project_with_status['fly_status'] = fly_status
         
-        logger.info(f"✅ Returning project details for: {project['name']}")
-        return jsonify({
-            'project': project_with_status
-        })
-        
+        return jsonify(project_with_status)
     except Exception as e:
         logger.error(f"💥 Error fetching project: {str(e)}")
         return jsonify({'error': 'Failed to fetch project'}), 500
 
 @projects_bp.route('/api/projects', methods=['POST'])
 def create_project():
-    """Create a new project"""
+    """Create a new project with GitHub repo and Fly.io app"""
     logger.info("🆕 POST /api/projects - Creating new project")
-    logger.debug(f"📥 Request headers: {dict(request.headers)}")
-    logger.debug(f"📥 Request remote addr: {request.remote_addr}")
-    logger.debug(f"📥 Request content type: {request.content_type}")
     
     try:
         # Get UUID from headers
         user_uuid = request.headers.get('X-User-UUID')
-        logger.debug(f"🆔 Raw UUID from header: '{user_uuid}'")
-        
         if not user_uuid:
             logger.warning("❌ X-User-UUID header is required")
-            logger.debug(f"📋 Available headers: {list(request.headers.keys())}")
             return jsonify({'error': 'X-User-UUID header is required'}), 400
         
         user_uuid = user_uuid.strip()
-        logger.debug(f"🆔 Cleaned UUID: '{user_uuid}' (length: {len(user_uuid)})")
-        
         if not user_uuid:
             logger.warning("❌ Empty UUID provided in header")
             return jsonify({'error': 'UUID cannot be empty'}), 400
         
         # Get request data
         data = request.get_json()
-        logger.debug(f"📥 Request data: {data}")
-        
         if not data or 'name' not in data:
             logger.warning("❌ Project name is required")
-            logger.debug(f"📥 Request data keys: {list(data.keys()) if data else 'None'}")
             return jsonify({'error': 'Project name is required'}), 400
         
         project_name = data['name'].strip()
-        logger.debug(f"📝 Project name: '{project_name}' (length: {len(project_name)})")
-        
         if not project_name:
             logger.warning("❌ Project name cannot be empty")
             return jsonify({'error': 'Project name cannot be empty'}), 400
         
-        # Create slug from name
-        slug = create_slug(project_name)
-        if not slug:
+        # Create slug from name (use provided slug if available, otherwise generate)
+        project_slug = data.get('slug', create_slug(project_name)).strip()
+        if not project_slug:
+            project_slug = create_slug(project_name)
+        
+        if not project_slug:
             logger.warning("❌ Invalid project name - cannot create slug")
             return jsonify({'error': 'Invalid project name'}), 400
         
-        logger.info(f"🔄 Creating project: {project_name} -> {slug}")
+        logger.info(f"🔄 Creating project: {project_name} -> {project_slug}")
         
         # Load existing projects
         projects = load_projects()
         
         # Check if project with same slug already exists
-        existing_project = next((p for p in projects if p.get('slug') == slug), None)
+        existing_project = next((p for p in projects if p.get('slug') == project_slug), None)
         if existing_project:
-            logger.warning(f"❌ Project with slug '{slug}' already exists")
+            logger.warning(f"❌ Project with slug '{project_slug}' already exists")
             return jsonify({'error': f'Project with name "{project_name}" already exists'}), 409
         
-        # Get user's API keys for GitHub and Fly.io
+        # Get user's API keys
         user_keys = load_user_keys(user_uuid)
         github_token = user_keys.get('github')
         fly_token = user_keys.get('fly')
         
         if not github_token:
-            logger.warning(f"❌ GitHub API key not found for user {user_uuid[:8]}")
-            return jsonify({'error': 'GitHub API key is required. Please set it up in integrations.'}), 400
+            logger.warning("❌ GitHub API key is required")
+            return jsonify({'error': 'GitHub API key is required. Please set it in integrations first.'}), 400
         
-        # Create Fly.io app first (only if Fly.io token is provided)
-        fly_app_data = None
-        if fly_token:
-            logger.info(f"🛩️ Creating Fly.io app: {slug}")
-            success, result = create_fly_app(slug, fly_token)
-            
-            if not success:
-                logger.error(f"❌ Fly.io app creation failed: {result}")
-                return jsonify({'error': result}), 409
-            
-            fly_app_data = result
-            logger.info(f"✅ Fly.io app created successfully: {slug}")
-        else:
-            logger.warning(f"⚠️ No Fly.io token provided - skipping Fly.io app creation")
+        if not fly_token:
+            logger.warning("❌ Fly.io API key is required")
+            return jsonify({'error': 'Fly.io API key is required. Please set it in integrations first.'}), 400
         
         # Create GitHub repository
-        success, result = create_github_repo(slug, github_token, fly_token)
-        if not success:
-            logger.error(f"❌ Failed to create GitHub repo: {result}")
-            
-            # If we created a Fly.io app but GitHub repo creation failed, 
-            # we should note this in the error but not delete the Fly.io app
-            # as the user might want to use it manually
-            if fly_app_data:
-                logger.warning(f"⚠️ Fly.io app '{slug}' was created but GitHub repo creation failed")
-                error_msg = f"GitHub repository creation failed: {result}. Note: Fly.io app '{slug}' was created successfully and can be used manually."
-                return jsonify({'error': error_msg}), 500
-            
-            return jsonify({'error': result}), 500
+        logger.info(f"🐙 Creating GitHub repository: {project_slug}")
+        github_success, github_result = create_github_repo(project_slug, github_token, fly_token)
+        
+        if not github_success:
+            logger.error(f"❌ GitHub repo creation failed: {github_result}")
+            return jsonify({'error': f'Failed to create GitHub repository: {github_result}'}), 500
+        
+        github_url = github_result
+        logger.info(f"✅ GitHub repository created: {github_url}")
+        
+        # Create Fly.io app
+        logger.info(f"🛩️ Creating Fly.io app: {project_slug}")
+        fly_success, fly_result = create_fly_app(project_slug, fly_token)
+        
+        if not fly_success:
+            logger.error(f"❌ Fly.io app creation failed: {fly_result}")
+            # Don't fail the entire project creation if Fly.io fails
+            # The user can manually create the app later
+            logger.warning(f"⚠️ Continuing with project creation despite Fly.io failure")
+            fly_app_name = None
+        else:
+            fly_app_name = project_slug
+            logger.info(f"✅ Fly.io app created: {fly_app_name}")
         
         # Create project record
         project = {
             'id': len(projects) + 1,
             'name': project_name,
-            'slug': slug,
-            'github_url': result,
+            'slug': project_slug,
+            'github_url': github_url,
+            'fly_app_name': fly_app_name,
             'created_at': datetime.utcnow().isoformat(),
             'created_by': user_uuid
         }
-        
-        # Add Fly.io app information if available
-        if fly_app_data:
-            project['fly_app'] = {
-                'id': fly_app_data.get('id'),
-                'name': slug,
-                'organization': fly_app_data.get('organization', {}).get('slug'),
-                'created_at': fly_app_data.get('created_at')
-            }
         
         # Add to projects list
         projects.append(project)
@@ -934,205 +852,90 @@ def create_project():
         logger.info(f"✅ Project created successfully: {project_name}")
         return jsonify({
             'message': 'Project created successfully',
-            'project': project
+            'project': project,
+            'warnings': [] if fly_success else [f'Fly.io app creation failed: {fly_result}']
         }), 201
         
     except Exception as e:
         logger.error(f"💥 Error creating project: {str(e)}")
         return jsonify({'error': 'Failed to create project'}), 500
 
-@projects_bp.route('/api/projects/<int:project_id>/conversations', methods=['GET'])
-def get_conversations(project_id):
-    """Get all conversations for a specific project"""
-    logger.info(f"📋 GET /api/projects/{project_id}/conversations - Fetching conversations")
-    
-    try:
-        # Verify project exists
-        projects = load_projects()
-        project = next((p for p in projects if p.get('id') == project_id), None)
-        if not project:
-            logger.warning(f"❌ Project not found: {project_id}")
-            return jsonify({'error': 'Project not found'}), 404
-        
-        conversations = load_conversations(project_id)
-        # Sort conversations by creation date (newest first)
-        conversations.sort(key=lambda x: x.get('created_at', ''), reverse=True)
-        
-        logger.info(f"📊 Returning {len(conversations)} conversations for project {project_id}")
-        return jsonify({
-            'conversations': conversations,
-            'count': len(conversations),
-            'project_id': project_id
-        })
-    except Exception as e:
-        logger.error(f"💥 Error fetching conversations: {str(e)}")
-        return jsonify({'error': 'Failed to fetch conversations'}), 500
-
-@projects_bp.route('/api/projects/<int:project_id>/conversations', methods=['POST'])
-def create_conversation(project_id):
-    """Create a new conversation for a specific project"""
-    logger.info(f"🆕 POST /api/projects/{project_id}/conversations - Creating new conversation")
-    
-    try:
-        # Get UUID from headers
-        user_uuid = request.headers.get('X-User-UUID')
-        if not user_uuid:
-            logger.warning("❌ X-User-UUID header is required")
-            return jsonify({'error': 'X-User-UUID header is required'}), 400
-        
-        user_uuid = user_uuid.strip()
-        if not user_uuid:
-            logger.warning("❌ Empty UUID provided in header")
-            return jsonify({'error': 'UUID cannot be empty'}), 400
-        
-        # Verify project exists
-        projects = load_projects()
-        project = next((p for p in projects if p.get('id') == project_id), None)
-        if not project:
-            logger.warning(f"❌ Project not found: {project_id}")
-            return jsonify({'error': 'Project not found'}), 404
-        
-        # Get request data
-        data = request.get_json()
-        if not data or 'name' not in data:
-            logger.warning("❌ Conversation name is required")
-            return jsonify({'error': 'Conversation name is required'}), 400
-        
-        conversation_name = data['name'].strip()
-        if not conversation_name:
-            logger.warning("❌ Conversation name cannot be empty")
-            return jsonify({'error': 'Conversation name cannot be empty'}), 400
-        
-        # Create slug from name (use provided slug if available, otherwise generate)
-        conversation_slug = data.get('slug', create_slug(conversation_name)).strip()
-        if not conversation_slug:
-            conversation_slug = create_slug(conversation_name)
-        
-        if not conversation_slug:
-            logger.warning("❌ Invalid conversation name - cannot create slug")
-            return jsonify({'error': 'Invalid conversation name'}), 400
-        
-        logger.info(f"🔄 Creating conversation: {conversation_name} -> {conversation_slug}")
-        
-        # Load existing conversations
-        conversations = load_conversations(project_id)
-        
-        # Check if conversation with same slug already exists
-        existing_conversation = next((c for c in conversations if c.get('slug') == conversation_slug), None)
-        if existing_conversation:
-            logger.warning(f"❌ Conversation with slug '{conversation_slug}' already exists")
-            return jsonify({'error': f'Conversation with name "{conversation_name}" already exists'}), 409
-        
-        # Create conversation record
-        conversation = {
-            'id': len(conversations) + 1,
-            'name': conversation_name,
-            'slug': conversation_slug,
-            'project_id': project_id,
-            'created_at': datetime.utcnow().isoformat(),
-            'created_by': user_uuid,
-            'last_message_at': None,
-            'message_count': 0
-        }
-        
-        # Add to conversations list
-        conversations.append(conversation)
-        
-        # Save to file
-        if not save_conversations(project_id, conversations):
-            logger.error("❌ Failed to save conversation to file")
-            return jsonify({'error': 'Failed to save conversation'}), 500
-        
-        logger.info(f"✅ Conversation created successfully: {conversation_name}")
-        return jsonify({
-            'message': 'Conversation created successfully',
-            'conversation': conversation
-        }), 201
-        
-    except Exception as e:
-        logger.error(f"💥 Error creating conversation: {str(e)}")
-        return jsonify({'error': 'Failed to create conversation'}), 500
-
-
 @projects_bp.route('/api/projects/<slug>', methods=['DELETE'])
 def delete_project(slug):
     """Delete a project and its associated GitHub repo and Fly.io app"""
     logger.info(f"🗑️ DELETE /api/projects/{slug} - Deleting project")
-
+    
     try:
         # Get UUID from headers
         user_uuid = request.headers.get('X-User-UUID')
         if not user_uuid:
             logger.warning("❌ X-User-UUID header is required")
             return jsonify({'error': 'X-User-UUID header is required'}), 400
-
+        
         user_uuid = user_uuid.strip()
         if not user_uuid:
             logger.warning("❌ Empty UUID provided in header")
             return jsonify({'error': 'UUID cannot be empty'}), 400
-
+        
         # Load projects
         projects = load_projects()
         
-        # Find the project to delete
+        # Find project by slug
         project = next((p for p in projects if p.get('slug') == slug), None)
         if not project:
             logger.warning(f"❌ Project not found: {slug}")
             return jsonify({'error': 'Project not found'}), 404
-
-        logger.info(f"🔍 Found project to delete: {project.get('name')} ({slug})")
-
-        # Load user keys for API operations
+        
+        logger.info(f"🔍 Found project to delete: {project['name']}")
+        
+        # Get user's API keys
         user_keys = load_user_keys(user_uuid)
         github_token = user_keys.get('github')
         fly_token = user_keys.get('fly')
-
+        
         deletion_results = {
-            'project_deleted': False,
-            'github_deleted': False,
-            'fly_deleted': False,
+            'github_success': False,
             'github_error': None,
+            'fly_success': False,
             'fly_error': None
         }
-
-        # Delete GitHub repository if it exists
-        github_url = project.get('github_url')
-        if github_url and github_token:
-            logger.info(f"🗑️ Deleting GitHub repository: {github_url}")
-            github_success, github_message = delete_github_repo(github_url, github_token)
-            deletion_results['github_deleted'] = github_success
+        
+        # Delete GitHub repository if URL exists and token is available
+        if project.get('github_url') and github_token:
+            logger.info(f"🐙 Deleting GitHub repository: {project['github_url']}")
+            github_success, github_message = delete_github_repo(project['github_url'], github_token)
+            deletion_results['github_success'] = github_success
             if not github_success:
                 deletion_results['github_error'] = github_message
                 logger.warning(f"⚠️ GitHub deletion failed: {github_message}")
-        elif github_url and not github_token:
-            deletion_results['github_error'] = "GitHub token not available"
-            logger.warning("⚠️ GitHub token not available for deletion")
+            else:
+                logger.info(f"✅ GitHub repository deleted: {github_message}")
         else:
-            logger.info("ℹ️ No GitHub repository to delete")
-
-        # Delete Fly.io app if it exists
-        if fly_token:
-            logger.info(f"🗑️ Deleting Fly.io app: {slug}")
-            fly_success, fly_message = delete_fly_app(slug, fly_token)
-            deletion_results['fly_deleted'] = fly_success
+            logger.info("⚠️ Skipping GitHub deletion (no URL or token)")
+        
+        # Delete Fly.io app if name exists and token is available
+        if project.get('fly_app_name') and fly_token:
+            logger.info(f"🛩️ Deleting Fly.io app: {project['fly_app_name']}")
+            fly_success, fly_message = delete_fly_app(project['fly_app_name'], fly_token)
+            deletion_results['fly_success'] = fly_success
             if not fly_success:
                 deletion_results['fly_error'] = fly_message
                 logger.warning(f"⚠️ Fly.io deletion failed: {fly_message}")
+            else:
+                logger.info(f"✅ Fly.io app deleted: {fly_message}")
         else:
-            deletion_results['fly_error'] = "Fly.io token not available"
-            logger.warning("⚠️ Fly.io token not available for deletion")
-
-        # Remove project from the list
+            logger.info("⚠️ Skipping Fly.io deletion (no app name or token)")
+        
+        # Remove project from list
         projects = [p for p in projects if p.get('slug') != slug]
         
         # Save updated projects list
         if save_projects(projects):
-            deletion_results['project_deleted'] = True
             logger.info(f"✅ Project {slug} removed from database")
         else:
             logger.error(f"❌ Failed to save projects after deletion")
             return jsonify({'error': 'Failed to update project database'}), 500
-
+        
         # Delete associated conversations file
         try:
             conversations_file = get_conversations_file(project.get('id'))
@@ -1141,7 +944,7 @@ def delete_project(slug):
                 logger.info(f"✅ Deleted conversations file for project {slug}")
         except Exception as e:
             logger.warning(f"⚠️ Failed to delete conversations file: {str(e)}")
-
+        
         # Prepare response
         response_data = {
             'message': f'Project "{project.get("name")}" deleted successfully',
@@ -1149,7 +952,7 @@ def delete_project(slug):
             'project_slug': slug,
             'deletion_results': deletion_results
         }
-
+        
         # Add warnings if some deletions failed
         warnings = []
         if deletion_results['github_error']:
@@ -1159,10 +962,10 @@ def delete_project(slug):
         
         if warnings:
             response_data['warnings'] = warnings
-
+        
         logger.info(f"✅ Project deletion completed: {slug}")
-        return jsonify(response_data), 200
-
+        return jsonify(response_data)
+        
     except Exception as e:
         logger.error(f"💥 Error deleting project: {str(e)}")
         return jsonify({'error': 'Failed to delete project'}), 500
